backend/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
fastapi==0.115.12 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.24.2 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.39.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-firestore==2.20.2 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-secret-manager==2.23.3 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.70.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.14.2 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.71.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.71.0 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.29.4 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.33.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.3 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.46.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.4.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.34.1 ; python_version >= "3.11" and python_version < "4.0"
firebase-admin>=6.0,<7.0
google-cloud-firestore>=2.7,<3.0
tenacity>=8.0.0,<9.0  # For retry logic with exponential backoff
